Upgrade data type for sizes, lengths, offsets, etc to 64-bit `long`

+++ Runtime/Scripts/GLTFSceneImporter.cs
@@ -1531,7 +1531,7 @@ namespace UnityGLTF
 			return new BufferCacheData
 			{
 				Stream = _gltfStream.Stream,
-				ChunkDataOffset = (uint)_gltfStream.Stream.Position,
+				ChunkDataOffset = _gltfStream.Stream.Position,
 				bufferData = GetOrCreateNativeBuffer(_gltfStream.Stream),
 			};
 		}
